% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queensReport.R
\name{queensReport}
\alias{queensReport}
\title{Queens Report (PDF)}
\usage{
queensReport(template = "QueensTemplate", resetStyleFiles = FALSE, ...)
}
\arguments{
\item{template}{character; Pandoc template to use for rendering. Pass
\code{"QueensTemplate"} to use the default example template}

\item{resetStyleFiles}{logical; should the style files (logo, cover, defs.tex) be overwritten with the default files?}

\item{...}{further arguments passed to \code{\link[rmarkdown]{pdf_document}}}
}
\description{
This function can be used to render the Queens report template. This is based on https://github.com/INWTlab/ireports
}
\details{
The function serves as wrapper to \code{\link[rmarkdown]{pdf_document}}
only steering the selection of the template.
}
